fix issues with create-modular-react-app (#492)
