+++ .changeset/four-ligers-laugh.md
@@ -1,0 +1,5 @@
+---
+'create-modular-react-app': patch
+---
+
+Fix issue with creating a new modular app with a name
