+++ packages/create-modular-react-app/.gitignore
@@ -1,0 +1,1 @@
+new-modular-app
