+++ packages/create-modular-react-app/package.json
@@ -9,6 +9,7 @@
     "node": ">=12"
   },
   "scripts": {
+    "create-modular-react-app": "ts-node src/cli.ts",
     "clean": "rimraf build",
     "prebuild": "yarn clean",
     "build": "babel --root-mode upward src --out-dir build --extensions .ts --ignore 'src/**/*.test.ts'"
